Move to package in family. Remove dependency on GiveBloc. Remove unused code.
